@@ -1,6 +1,6 @@
 name: jiffy
 description: Jiffy is a date dart package inspired by momentjs for parsing, manipulating and formatting dates
-version: 3.0.1
+version: 4.0.0
 homepage: https://github.com/jama5262/jiffy
 repository: https://github.com/jama5262/jiffy
 issue_tracker: https://github.com/jama5262/jiffy/issues
@@ -8,13 +8,11 @@ documentation: https://github.com/jama5262/jiffy/tree/master/doc
 author: Jama Mohamed <jama3137@gmail.com>
 
 environment:
-  sdk: '>=2.4.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  intl: ^0.16.0
+  intl: ^0.17.0
 
 dev_dependencies:
-  pedantic: ^1.7.0
-  test: ^1.6.0
-  flutter_test:
-    sdk: flutter
+  pedantic: ^1.10.0
+  test: ^1.16.5
