@@ -2,12 +2,14 @@ import 'package:jiffy/src/relative_time/locales/ar_locale.dart';
 import 'package:jiffy/src/relative_time/locales/de_locale.dart';
 import 'package:jiffy/src/relative_time/locales/en_locale.dart';
 import 'package:jiffy/src/relative_time/locales/es_locale.dart';
+import 'package:jiffy/src/relative_time/locales/fa_locale.dart';
 import 'package:jiffy/src/relative_time/locales/fr_locale.dart';
 import 'package:jiffy/src/relative_time/locales/hi_locale.dart';
 import 'package:jiffy/src/relative_time/locales/id_locale.dart';
 import 'package:jiffy/src/relative_time/locales/it_locale.dart';
 import 'package:jiffy/src/relative_time/locales/ja_locale.dart';
 import 'package:jiffy/src/relative_time/locales/ko_locale.dart';
+import 'package:jiffy/src/relative_time/locales/nb_locale.dart';
 import 'package:jiffy/src/relative_time/locales/pl_locale.dart';
 import 'package:jiffy/src/relative_time/locales/pt_locale.dart';
 import 'package:jiffy/src/relative_time/locales/ru_locale.dart';
@@ -36,6 +38,7 @@ Map<String, LookUpMessages> _lookupMessagesMap = {
   'zhtw': ZhLocale(),
   'ja': JaLocale(),
   'de': DeLocale(),
+  'dede': DeLocale(),
   'deat': DeLocale(),
   'dech': DeLocale(),
   'fr': FrLocale(),
@@ -57,10 +60,12 @@ Map<String, LookUpMessages> _lookupMessagesMap = {
   'pl': PlLocale(),
   'ptbr': PtLocale(),
   'tr': TrLocale(),
-  'sv': SvLocale()
+  'sv': SvLocale(),
+  'nb': NbLocale(),
+  'fa': FaLocale(true),
 };
 
-String format(String locale, DateTime date1, [DateTime date2]) {
+String format(String locale, DateTime date1, [DateTime? date2]) {
   final messages =
       _lookupMessagesMap[replaceLocaleHyphen(locale)] ?? EnLocale();
   final _date2 = date2 ?? DateTime.now();
@@ -111,6 +116,6 @@ String format(String locale, DateTime date1, [DateTime date2]) {
   }
 
   return [prefix, result, suffix]
-      .where((str) => str != null && str.isNotEmpty)
+      .where((str) => str.isNotEmpty)
       .join(messages.wordSeparator());
 }
