@@ -1,4 +1,4 @@
-List<String> localeOrdinals = [
+List<String> _localeOrdinals = [
   'en',
   'es',
   'fr',
@@ -17,34 +17,30 @@ List<String> localeOrdinals = [
   'de',
   'deat',
   'dech',
-  'id',
-  'tr',
-  'sv'
+  'tr'
 ];
 
-List<String> getOrdinalLocaleDates(String locale) {
+List<String>? getOrdinalLocaleDates(String locale) {
   var ordinals = {
-    localeOrdinals[0]: ['th', 'st', 'nd', 'rd'],
-    localeOrdinals[1]: ['º', 'º', 'º', 'º'],
-    localeOrdinals[2]: ['', 'er', '', ''],
-    localeOrdinals[3]: ['e', 'er', 'e', 'e'],
-    localeOrdinals[4]: ['e', 'er', 'e', 'e'],
-    localeOrdinals[5]: ['º', 'º', 'º', 'º'],
-    localeOrdinals[6]: ['º', 'º', 'º', 'º'],
-    localeOrdinals[7]: ['日', '日', '日', '日'],
-    localeOrdinals[8]: ['일', '일', '일', '일'],
-    localeOrdinals[9]: ['º', 'º', 'º', 'º'],
-    localeOrdinals[10]: ['º', 'º', 'º', 'º'],
-    localeOrdinals[11]: ['日', '日', '日', '日'],
-    localeOrdinals[12]: ['日', '日', '日', '日'],
-    localeOrdinals[13]: ['日', '日', '日', '日'],
-    localeOrdinals[14]: ['日', '日', '日', '日'],
-    localeOrdinals[15]: ['.', '.', '.', '.'],
-    localeOrdinals[16]: ['.', '.', '.', '.'],
-    localeOrdinals[17]: ['.', '.', '.', '.'],
-    localeOrdinals[18]: ['', '', '', ''],
-    localeOrdinals[19]: ['.', '.', '.', '.'],
-    localeOrdinals[20]: ['', '', '', ''],
+    _localeOrdinals[0]: ['th', 'st', 'nd', 'rd'],
+    _localeOrdinals[1]: ['º', 'º', 'º', 'º'],
+    _localeOrdinals[2]: ['', 'er', '', ''],
+    _localeOrdinals[3]: ['e', 'er', 'e', 'e'],
+    _localeOrdinals[4]: ['e', 'er', 'e', 'e'],
+    _localeOrdinals[5]: ['º', 'º', 'º', 'º'],
+    _localeOrdinals[6]: ['º', 'º', 'º', 'º'],
+    _localeOrdinals[7]: ['日', '日', '日', '日'],
+    _localeOrdinals[8]: ['일', '일', '일', '일'],
+    _localeOrdinals[9]: ['º', 'º', 'º', 'º'],
+    _localeOrdinals[10]: ['º', 'º', 'º', 'º'],
+    _localeOrdinals[11]: ['日', '日', '日', '日'],
+    _localeOrdinals[12]: ['日', '日', '日', '日'],
+    _localeOrdinals[13]: ['日', '日', '日', '日'],
+    _localeOrdinals[14]: ['日', '日', '日', '日'],
+    _localeOrdinals[15]: ['.', '.', '.', '.'],
+    _localeOrdinals[16]: ['.', '.', '.', '.'],
+    _localeOrdinals[17]: ['.', '.', '.', '.'],
+    _localeOrdinals[18]: ['.', '.', '.', '.']
   };
   return ordinals[locale];
 }
